Add general file viewer and annotator
This commit adds a FileViewer component, which handles the logic of
deciding whether a PDF or a code file should be displayed. This
component can replace the PDFViewer with an extra prop (assignment
type) added.
This commit also adds a FileAnnotator component, which does the same as
the FileViewer, but decides between a PDFAnnotator and a CodeViewer (
for now).