Integrate file tree into code viewer
This replaces the CodeViewer with a CodeWrapper to integrate the
FileTree in a code duplication preventing manner. The CodeViewer is now
only responsible for rendering content. The CodeAnnotator will be
inserted into the CodeWrapper when it is done, thus preventing the
need for duplicating file loading and the file tree.