Add file verification
This commit adds a relatively simple check for certain characters in a
file. If it finds those characters, a warning will be displayed to the
user after which they can decide to view the file anyway or not.

This commit also adds a simple loading overlay to make it more clear
when a new file is being loaded / prepared.