Fix merge request threads
This commit fixes the following things that were brought up in the merge
request.
 - The loading and warning overlay now have a maximum height of 80% of
   the window height.
 - The file verification is now also applied to single file submissions

During testing I also noticed that the line number implementation in the
CodeViewer lead to a scrollbar in the gutter. This has been fixed by
applying the same layout and styling to the CodeViewer as was present in
the CodeAnnotations.