Fix multi line highlighting
This commit fixes the problem where things like multi line comments were
not highlighted correctly.

This fix has been implemented using a stack.
I thought of using a regex expression to automatically do what the for-
loop and stack do, but that expression would be far too complex (if it
is even possible).