Implement annotation aggregation
The method that has been implemented in this commit creates a new
annotation for each overlap.
For this method it helps to see an annotation as having a set of line
numbers over which it annotates.
Assume A and B both include lines 6-8, then three comments will be
created. A - B with the text of A, B - A with the text of B and A & B
with the text of A and B.
Duplicates are discarded, so if A and B cover exactly the same lines,
we only get A and B.